pyproject.toml
[project]
name = "wa-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "openai>=1.106.1",
    "openai-agents>=0.2.11",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "twilio>=9.8.0",
    "uvicorn>=0.35.0",
]
